+++ CHANGELOG.md
@@ -69,7 +69,8 @@ All notable changes to this project will be documented in this file.
   - `plus(String)`, `plus(NotBlankString)` and `plus(Char)` operations on
     `NotBlankString`
   - `Char.plus(NotBlankString)` operation
-  - `StrictlyPositiveInt.Companion.range` property.
+  - `range` property for companion object of `StrictlyPositiveInt` and
+    `StrictlyNegativeInt`.
 - Opt-in message of experimental annotations (pull request
   [#328](https://github.com/kotools/types/pull/328)).
 - Documentation of types in API reference and in README (commit